Stage 4 completed.

+++ src/main/java/org/example/services/impl/UserServiceImpl.java
@@ -3,17 +3,19 @@ package org.example.services.impl;
 import lombok.RequiredArgsConstructor;
 import org.example.models.entities.UserEntity;
 import org.example.models.enums.RoleEnum;
-import org.example.models.requests.ChangeUserAccessRequest;
-import org.example.models.requests.ChangeUserRoleRequest;
-import org.example.models.requests.RegisterUserRequest;
 import org.example.models.responses.ChangeUserAccessResponse;
+import org.example.models.responses.ChangeUserRoleResponse;
 import org.example.models.responses.DeleteUserResponse;
 import org.example.models.responses.RegisterUserResponse;
 import org.example.repositories.UserRepository;
 import org.example.services.UserService;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.Optional;
@@ -26,9 +28,12 @@ public class UserServiceImpl implements UserService {
     final PasswordEncoder passwordEncoder;
 
     @Override
-    public RegisterUserResponse register(RegisterUserRequest request) {
-        UserEntity userEntity = new UserEntity(request.getName(), request.getUsername());
-        userEntity.setPassword(passwordEncoder.encode(request.getPassword()));
+    public RegisterUserResponse register(String name, String username, String password) {
+        if (usernameAlreadyExists(username)) {
+            throw new ResponseStatusException(HttpStatus.CONFLICT);
+        }
+        UserEntity userEntity = new UserEntity(name, username);
+        userEntity.setPassword(passwordEncoder.encode(password));
 
         if (userRepository.count() >= 1) {
             userEntity.setRole(RoleEnum.MERCHANT);
@@ -60,6 +65,9 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public DeleteUserResponse delete(String username) {
+        if (findUserByUsername(username).isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        }
         UserEntity user = findUserByUsername(username)
                 .orElseThrow(() -> new UsernameNotFoundException("Username " + username + " doesn't exist!"));
         userRepository.delete(user);
@@ -72,13 +80,26 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public void changeRole(ChangeUserRoleRequest request) {
-        Optional<UserEntity> userEntity = findUserByUsername(request.getUsername());
-        if (userEntity.isPresent()) {
-            RoleEnum role = RoleEnum.valueOf(request.getRole());
-            userEntity.get().setRole(role);
-            userRepository.save(userEntity.get());
+    public ChangeUserRoleResponse changeRole(String username, String userRole) {
+        if (!checkIfRoleExist(userRole)) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
         }
+        if (userRole.equals(RoleEnum.ADMINISTRATOR.name())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+        }
+
+        Optional<UserEntity> user = findUserByUsername(username);
+        if (user.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        }
+        if (userRole.equals(user.get().getRole().name())) {
+            throw new ResponseStatusException(HttpStatus.CONFLICT);
+        }
+        RoleEnum role = RoleEnum.valueOf(userRole);
+        user.get().setRole(role);
+        userRepository.save(user.get());
+        return new ChangeUserRoleResponse(user.get().getId(), user.get().getName(),
+                                          user.get().getUsername(), user.get().getRole().toString());
     }
 
     @Override
@@ -89,32 +110,39 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public ChangeUserAccessResponse changeAccess(ChangeUserAccessRequest request) {
-        ChangeUserAccessResponse changeUserAccessResponse;
-        if (request.getOperation().equals("UNLOCK")) {
-            changeUserAccessResponse = unlockUser(request);
-        } else {
-            changeUserAccessResponse = lockUser(request);
+    public ChangeUserAccessResponse changeAccess(String username, String operation) {
+        ChangeUserAccessResponse changeUserAccessResponse = null;
+        if (operation.equals("UNLOCK")) {
+            changeUserAccessResponse = unlockUser(username);
+        } else if (operation.equals("LOCK")) {
+            changeUserAccessResponse = lockUser(username);
         }
         return changeUserAccessResponse;
     }
 
-    private ChangeUserAccessResponse unlockUser(ChangeUserAccessRequest request) {
-        UserEntity userEntity = findUserByUsername(request.getUsername())
+    @Override
+    public Optional<UserEntity> getLoggedUser() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String username = authentication.getName();
+        return findUserByUsername(username);
+    }
+
+    private ChangeUserAccessResponse unlockUser(String username) {
+        UserEntity userEntity = findUserByUsername(username)
                 .orElseThrow(
-                        () -> new UsernameNotFoundException("Username " + request.getUsername() + " doesn't exist!"));
+                        () -> new UsernameNotFoundException("Username " + username + " doesn't exist!"));
         userEntity.setAccountLocked(true);
         userRepository.save(userEntity);
-        return new ChangeUserAccessResponse("User " + request.getUsername() + " unlocked!");
+        return new ChangeUserAccessResponse("User " + username + " unlocked!");
     }
 
 
-    private ChangeUserAccessResponse lockUser(ChangeUserAccessRequest request) {
-        UserEntity userEntity = findUserByUsername(request.getUsername())
+    private ChangeUserAccessResponse lockUser(String username) {
+        UserEntity userEntity = findUserByUsername(username)
                 .orElseThrow(
-                        () -> new UsernameNotFoundException("Username " + request.getUsername() + " doesn't exist!"));
+                        () -> new UsernameNotFoundException("Username " + username + " doesn't exist!"));
         userEntity.setAccountLocked(false);
         userRepository.save(userEntity);
-        return new ChangeUserAccessResponse("User " + request.getUsername() + " locked!");
+        return new ChangeUserAccessResponse("User " + username + " locked!");
     }
 }
